refactored PlanPage into smaller components
