+++ src/pages/PlanPage/components/Main.tsx
@@ -1,0 +1,13 @@
+import {GetSelectedSubPage} from "./../components/GetSelectedSubPage";
+import {Box} from "@mui/material";
+
+
+function Main({selectedSubpage,selectedCareer, selectedCollege, setSelectedCareer, setSelectedCollege}:any) {
+  return (
+    <Box component='main' sx={{ flex: 1, py: 6, px: 4, bgcolor: "#eaeff1" }}>
+    {GetSelectedSubPage(selectedSubpage,selectedCareer, selectedCollege, setSelectedCareer, setSelectedCollege)}
+  </Box>
+  )
+}
+
+export default Main
